Reducing confusion.
Date for each chapter to be read for.
Book name of the chapters wanted to be read. 

Currently we have: Chapter X from no book.
LYH - No obvious connection to a specific book.

The only real way to figure out what to read is to find the book that have a historical chapter 13. Or maybe I'm missing something.